delete import-r-lib.spec
